ENH: Add encryption support.
Adds new configurable ``crypto`` attributes to PostgresContentsManager
and PostgresCheckpoints.  ``crypto`` objects must provide ``encrypt``
and ``decrypt`` methods, which will be called on data being written to
and read from the database, respectively.

The default ``crypto`` implementation is ``NoEncryption``, which returns
its inputs unchanged.

The built-in ``crypto`` implementation is ``FernetEncryption``, which
uses ``cryptography.fernet`` to implement symmetric-key encryption of
all notebooks and files in the database.