Fixing upsert of file
First, a query is executed to check whether the file
already exists. If not, it is inserted. If it does exist
the file is updated.

The current code was given issues because the transaction
was closed after the failure.

In the future it might be better to use a SQLAlchemy Session
which handles upserts under the hood.